Divide KatagamiList code
