+++ front/src/components/lv2/PaginationActions.js
@@ -1,0 +1,60 @@
+import React from 'react'
+import { IconButton } from '@material-ui/core'
+import { makeStyles } from '@material-ui/styles'
+import {
+  FirstPage,
+  KeyboardArrowRight,
+  KeyboardArrowLeft,
+  LastPage,
+} from '@material-ui/icons'
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+}))
+
+export default props => {
+  const { count, page, rowsPerPage, onChangePage } = props
+  const classes = useStyles()
+
+  const handleFirstPageButtonClick = event => {
+    onChangePage(event, 0)
+  }
+
+  const handleBackButtonClick = event => {
+    onChangePage(event, page - 1)
+  }
+
+  const handleNextButtonClick = event => {
+    onChangePage(event, page + 1)
+  }
+
+  const handleLastPageButtonClick = event => {
+    onChangePage(event, Math.ceil(count / rowsPerPage) - 1)
+  }
+
+  return (
+    <div className={classes.root}>
+      <IconButton onClick={handleFirstPageButtonClick} disabled={page === 0}>
+        <FirstPage />
+      </IconButton>
+      <IconButton onClick={handleBackButtonClick} disabled={page === 0}>
+        <KeyboardArrowLeft />
+      </IconButton>
+      <IconButton
+        onClick={handleNextButtonClick}
+        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
+      >
+        <KeyboardArrowRight />
+      </IconButton>
+      <IconButton
+        onClick={handleLastPageButtonClick}
+        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
+      >
+        <LastPage />
+      </IconButton>
+    </div>
+  )
+}
