+++ front/src/components/lv3/KatagamiList.js
@@ -4,21 +4,15 @@ import {
   TableHead,
   TableRow,
   TableCell,
-  TableBody,
   TableFooter,
-  IconButton,
   TablePagination,
 } from '@material-ui/core'
 import { makeStyles } from '@material-ui/styles'
 import { currentUser } from 'libs/auth'
 import { fetchKatagamis } from 'libs/api'
 import theme from 'libs/theme'
-import {
-  FirstPage,
-  KeyboardArrowRight,
-  KeyboardArrowLeft,
-  LastPage,
-} from '@material-ui/icons'
+import PaginationActions from 'components/lv2/PaginationActions'
+import KatagamiListBody from 'components/lv2/KatagamiListBody'
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -27,27 +21,11 @@ const useStyles = makeStyles(theme => ({
     justifyContent: 'space-between',
     overflow: 'hidden',
   },
-  tableHead: {
-    backgroundColor: theme.palette.primary.light,
-  },
   tableRow: {
-    '& *': {
-      fontWeight: 'normal',
-    },
+    '& *': { fontWeight: 'normal' },
   },
-  done: {
-    color: theme.palette.primary.main,
-  },
-  yet: {
-    color: theme.palette.secondary.main,
-  },
-  footer: {
-    marginTop: theme.spacing(20),
-  },
-  footerButtons: {
-    display: 'flex',
-    justifyContent: 'space-between',
-  },
+  header: { backgroundColor: theme.palette.primary.light },
+  footer: { marginTop: theme.spacing(20) },
 }))
 
 export default function() {
@@ -85,49 +63,6 @@ export default function() {
     })
   }, [isLatest])
 
-  const PaginationActions = props => {
-    const { count, page, rowsPerPage, onChangePage } = props
-
-    const handleFirstPageButtonClick = event => {
-      onChangePage(event, 0)
-    }
-
-    const handleBackButtonClick = event => {
-      onChangePage(event, page - 1)
-    }
-
-    const handleNextButtonClick = event => {
-      onChangePage(event, page + 1)
-    }
-
-    const handleLastPageButtonClick = event => {
-      onChangePage(event, Math.ceil(count / rowsPerPage) - 1)
-    }
-
-    return (
-      <div className={classes.footerButtons}>
-        <IconButton onClick={handleFirstPageButtonClick} disabled={page === 0}>
-          <FirstPage />
-        </IconButton>
-        <IconButton onClick={handleBackButtonClick} disabled={page === 0}>
-          <KeyboardArrowLeft />
-        </IconButton>
-        <IconButton
-          onClick={handleNextButtonClick}
-          disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        >
-          <KeyboardArrowRight />
-        </IconButton>
-        <IconButton
-          onClick={handleLastPageButtonClick}
-          disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        >
-          <LastPage />
-        </IconButton>
-      </div>
-    )
-  }
-
   const handleChangePage = (event, newPage) => {
     handlePaginate({ page: newPage, per: rowsPerPage })
     setPage(newPage)
@@ -143,36 +78,14 @@ export default function() {
     <div className={classes.root}>
       <Table>
         <TableHead>
-          <TableRow className={classes.tableHead}>
+          <TableRow className={classes.header}>
             <TableCell align="right">id</TableCell>
             <TableCell align="left">ファイル名</TableCell>
             <TableCell align="right">アノテーション件数</TableCell>
             <TableCell align="center">あなたの達成度</TableCell>
           </TableRow>
         </TableHead>
-        <TableBody>
-          {katagamis.map(katagami => (
-            <TableRow key={katagami.id} className={classes.tableRow}>
-              <TableCell align="right">{katagami.id}</TableCell>
-              <TableCell>{katagami.name}</TableCell>
-              <TableCell align="right">{katagami.annotation_num}</TableCell>
-              {katagami.done_by_current_user ? (
-                <TableCell align="center" className={classes.done}>
-                  完了
-                </TableCell>
-              ) : (
-                <TableCell align="center" className={classes.yet}>
-                  未達成
-                </TableCell>
-              )}
-            </TableRow>
-          ))}
-          {emptyRows > 0 && (
-            <TableRow style={{ height: 55 * emptyRows }}>
-              <TableCell colSpan={6} />
-            </TableRow>
-          )}
-        </TableBody>
+        <KatagamiListBody katagamis={katagamis} emptyRows={emptyRows} />
         <TableFooter className={classes.footer}>
           <TableRow className={classes.tableRow}>
             <TablePagination
